Move function commands under the app top-level namespace
