+++ packages/app/src/cli/commands/app/function/run.ts
@@ -1,16 +1,16 @@
-import {appFlags} from '../../flags.js'
-import {buildGraphqlTypes} from '../../services/function/build.js'
-import {loadExtensionsSpecifications} from '../../models/extensions/specifications.js'
-import {load as loadApp} from '../../models/app/loader.js'
-import {AppInterface} from '../../models/app/app.js'
+import {appFlags} from '../../../flags.js'
+import {runFunctionRunner} from '../../../services/function/build.js'
+import {AppInterface} from '../../../models/app/app.js'
+import {load as loadApp} from '../../../models/app/loader.js'
+import {loadExtensionsSpecifications} from '../../../models/extensions/specifications.js'
 import Command from '@shopify/cli-kit/node/base-command'
 import {globalFlags} from '@shopify/cli-kit/node/cli'
 import {resolvePath, cwd} from '@shopify/cli-kit/node/path'
 import {AbortError} from '@shopify/cli-kit/node/error'
-import {renderFatalError, renderSuccess} from '@shopify/cli-kit/node/ui'
+import {renderFatalError} from '@shopify/cli-kit/node/ui'
 
-export default class FunctionTypegen extends Command {
-  static description = 'Generate GraphQL types for your JavaScript function.'
+export default class FunctionRun extends Command {
+  static description = 'Run a Shopify function locally for testing.'
 
   static flags = {
     ...globalFlags,
@@ -18,7 +18,7 @@ export default class FunctionTypegen extends Command {
   }
 
   public async run() {
-    const {flags} = await this.parse(FunctionTypegen)
+    const {flags} = await this.parse(FunctionRun)
     const directory = flags.path ? resolvePath(flags.path) : cwd()
 
     const specifications = await loadExtensionsSpecifications(this.config)
@@ -26,8 +26,7 @@ export default class FunctionTypegen extends Command {
 
     const ourFunction = app.extensions.function.find((fun) => fun.directory === directory)
     if (ourFunction) {
-      await buildGraphqlTypes(ourFunction.directory, {stdout: process.stdout, stderr: process.stderr})
-      renderSuccess({headline: 'GraphQL types generated successfully.'})
+      await runFunctionRunner(ourFunction)
     } else {
       renderFatalError(new AbortError('You should run this command from the root of a function.'))
     }
